modularize types and resolvers
