+++ src/api/comment.api.js
@@ -2,8 +2,7 @@ import Comment from '../models/comment.model';
 import User from '../models/user.model';
 
 export default {
-
-  createComment : args =>
+  createComment: args =>
     Promise.all([
       User.findById(args.userId),
       Comment.create(args)
