+++ src/app.js
@@ -1,72 +1,16 @@
-import {
-  ApolloServer,
-  gql
-} from 'apollo-server';
+import { ApolloServer, makeExecutableSchema } from 'apollo-server';
 
 import { verifyConnection } from './lib/db';
 import associations from './models/associations';
-
-import postApi from './api/post.api';
-import commentsApi from './api/comment.api';
-
 import selectionHelper from './utils/selectionHelper';
 
-const typeDefs = gql`
-  type Post {
-    id: ID
-    title: String
-    user: User
-    content: String
-    comments: [Comment]
-  }
-
-  type User {
-    id: ID
-    name: String
-    posts: [Post]
-  }
-
-  type Comment {
-    id: ID
-    description: String
-    post: Post
-    user: User
-  }
-
-  type Query {
-    posts: [Post]
-    post(id: ID): Post
-  }
-
-  type Mutation {
-    createPost(title: String, content:String, userId: ID): Post
-    commentPost(description: String, postId: ID, userId: ID): Comment
-  }
-`;
-
-const resolvers = {
-  Query: {
-    posts: () => postApi.getPosts(),
-
-    post: (root, args, ctx, info) => {
-      return postApi.getPost(args.id, selectionHelper.extractSelection(info));
-    }
-  },
-
-
-  Mutation: {
-    createPost(root, args) {
-      return postApi.createPost(args);
-    },
-
-    commentPost(root, args) {
-      return commentsApi.createComment(args);
-    }
-  }
-};
+import typeDefs from './types';
+import resolvers from './resolvers';
 
 const start = () => {
-  const server = new ApolloServer({ typeDefs, resolvers });
+  const server = new ApolloServer({
+    schema: makeExecutableSchema({ typeDefs, resolvers})
+  });
 
   // setup database associations
   associations();
