+++ src/resolvers/comments.resolvers.js
@@ -1,0 +1,13 @@
+import commentsApi from '../api/comment.api';
+
+export default {
+
+  queries: {
+
+  },
+
+  mutations: {
+    commentPost: (root, args) => {console.log('here'); return commentsApi.createComment(args)}
+  }
+
+}
