+++ src/resolvers/index.js
@@ -1,0 +1,14 @@
+import commentsResolvers from './comments.resolvers';
+import postsResolvers from './posts.resolvers';
+
+export default {
+  Query: {
+    ...postsResolvers.queries,
+    ...commentsResolvers.queries
+  },
+
+  Mutation: {
+    ...postsResolvers.mutations,
+    ...commentsResolvers.mutations
+  }
+}
