+++ src/types/comment.type.js
@@ -1,0 +1,8 @@
+export default `
+  type Comment {
+    id: ID
+    description: String
+    post: Post
+    user: User
+  }
+`;
