+++ src/types/post.type.js
@@ -1,0 +1,9 @@
+export default `
+  type Post {
+    id: ID
+    title: String
+    user: User
+    content: String
+    comments: [Comment]
+  }
+`;
