+++ src/types/user.type.js
@@ -1,0 +1,7 @@
+export default `
+  type User {
+    id: ID
+    name: String
+    posts: [Post]
+  }
+`;
